+++ docs/.vuepress/components/PerformanceNowExample.vue
@@ -1,0 +1,21 @@
+<template>
+  <div>
+    <p>
+      Performance.now : {{ now }}</b>
+    </p>
+    <p>
+      Created <b>{{Math.floor(now / 1000) }}</b> seconds ago
+    </p>
+  </div>
+</template>
+
+<script>
+import { usePerformanceNow } from "vue-composable";
+
+export default {
+  name: "performance-now-example",
+  setup() {
+    return usePerformanceNow();
+  }
+};
+</script>