+++ packages/core/src/date/index.ts
@@ -1,0 +1,3 @@
+export * from './dateNow'
+export * from './now'
+export * from './performanceNow'